fix(parser): split owning-project resolution out of unit-extractor.ts
unit-extractor.ts had grown to 446 lines, over the 400-line module-size
budget with no allowlist row. Moves the containment-walk helper
(resolveOwningProjectId, its RelatedLookup interface, and the walk-depth
constant) into a new packages/parser/src/owning-project.ts, re-exported
from unit-extractor.ts so existing internal imports keep working
unchanged. No public export or behaviour change: unit extraction stays
in unit-extractor.ts (now 365 lines), the new module is 100 lines.
